// packages/sotast/src/generated/solomon.ts
export interface BoundingBox {
  /**
   * @generated from protobuf field: float x1 = 1;
   */
  x1: number;
  /**
   * @generated from protobuf field: float y1 = 2;
   */
  y1: number;
  /**
   * @generated from protobuf field: float x2 = 3;
   */
  x2: number;
  /**
   * @generated from protobuf field: float y2 = 4;
   */
  y2: number;
}
/**
 * @generated from protobuf message me.sylver.marver.solomon.Vector
 */
export interface Vector {
  /**
   * @generated from protobuf field: repeated float value = 1;
   */
  value: number[];
}
// ...
export interface OCR {
  /**
   * @generated from protobuf field: string text = 1;
   */
  text: string;
  /**
   * @generated from protobuf field: me.sylver.marver.solomon.BoundingBox bounding_box = 2;
   */
  bounding_box?: BoundingBox;
  /**
   * @generated from protobuf field: float confidence = 3;
   */
  confidence: number;
}
// ...
class BoundingBox$Type extends MessageType<BoundingBox> {
  constructor() {
    super('me.sylver.marver.solomon.BoundingBox', [
      { no: 1, name: 'x1', kind: 'scalar', T: 2 /*ScalarType.FLOAT*/ },
      { no: 2, name: 'y1', kind: 'scalar', T: 2 /*ScalarType.FLOAT*/ },
      { no: 3, name: 'x2', kind: 'scalar', T: 2 /*ScalarType.FLOAT*/ },
      { no: 4, name: 'y2', kind: 'scalar', T: 2 /*ScalarType.FLOAT*/ },
    ]);
  }
  create(value?: PartialMessage<BoundingBox>): BoundingBox {
    const message = { x1: 0, y1: 0, x2: 0, y2: 0 };
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
    if (value !== undefined) reflectionMergePartial<BoundingBox>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: BoundingBox,
  ): BoundingBox {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* float x1 */ 1:
          message.x1 = reader.float();
          break;
        case /* float y1 */ 2:
          message.y1 = reader.float();
          break;
        case /* float x2 */ 3:
          message.x2 = reader.float();
          break;
        case /* float y2 */ 4:
          message.y2 = reader.float();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: BoundingBox,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* float x1 = 1; */
    if (message.x1 !== 0) writer.tag(1, WireType.Bit32).float(message.x1);
    /* float y1 = 2; */
    if (message.y1 !== 0) writer.tag(2, WireType.Bit32).float(message.y1);
    /* float x2 = 3; */
    if (message.x2 !== 0) writer.tag(3, WireType.Bit32).float(message.x2);
    /* float y2 = 4; */
    if (message.y2 !== 0) writer.tag(4, WireType.Bit32).float(message.y2);
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message me.sylver.marver.solomon.BoundingBox
 */
export const BoundingBox = new BoundingBox$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Vector$Type extends MessageType<Vector> {
  constructor() {
    super('me.sylver.marver.solomon.Vector', [
      { no: 1, name: 'value', kind: 'scalar', repeat: 1 /*RepeatType.PACKED*/, T: 2 /*ScalarType.FLOAT*/ },
    ]);
  }
  create(value?: PartialMessage<Vector>): Vector {
    const message = { value: [] };
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
    if (value !== undefined) reflectionMergePartial<Vector>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: Vector,
  ): Vector {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* repeated float value */ 1:
          if (wireType === WireType.LengthDelimited)
            for (let e = reader.int32() + reader.pos; reader.pos < e; ) message.value.push(reader.float());
          else message.value.push(reader.float());
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
      }
    }
    return message;
  }
  internalBinaryWrite(message: Vector, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
    /* repeated float value = 1; */
    if (message.value.length) {
      writer.tag(1, WireType.LengthDelimited).fork();
      for (let i = 0; i < message.value.length; i++) writer.float(message.value[i]);
      writer.join();
    }
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message me.sylver.marver.solomon.Vector
 */
export const Vector = new Vector$Type();
// ...
class OCR$Type extends MessageType<OCR> {
  constructor() {
    super('me.sylver.marver.solomon.OCR', [
      { no: 1, name: 'text', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: 'bounding_box', kind: 'message', localName: 'bounding_box', T: () => BoundingBox },
      { no: 3, name: 'confidence', kind: 'scalar', T: 2 /*ScalarType.FLOAT*/ },
    ]);
  }
  create(value?: PartialMessage<OCR>): OCR {
    const message = { text: '', confidence: 0 };
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
    if (value !== undefined) reflectionMergePartial<OCR>(this, message, value);
    return message;
  }
  internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: OCR): OCR {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string text */ 1:
          message.text = reader.string();
          break;
        case /* me.sylver.marver.solomon.BoundingBox bounding_box */ 2:
          message.bounding_box = BoundingBox.internalBinaryRead(
            reader,
            reader.uint32(),
            options,
            message.bounding_box,
          );
          break;
        case /* float confidence */ 3:
          message.confidence = reader.float();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
      }
    }
    return message;
  }
  internalBinaryWrite(message: OCR, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
    /* string text = 1; */
    if (message.text !== '') writer.tag(1, WireType.LengthDelimited).string(message.text);
    /* me.sylver.marver.solomon.BoundingBox bounding_box = 2; */
    if (message.bounding_box)
      BoundingBox.internalBinaryWrite(
        message.bounding_box,
        writer.tag(2, WireType.LengthDelimited).fork(),
        options,
      ).join();
    /* float confidence = 3; */
    if (message.confidence !== 0) writer.tag(3, WireType.Bit32).float(message.confidence);
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message me.sylver.marver.solomon.OCR
 */
export const OCR = new OCR$Type();